UDP
We modified the UDP to use our own UDP object. This is different form
Matlab toolbox but should be able to take the same parameters and
operate the same way. The UDP sockets were modified to not close and
reopen the ports every time the remote sending port changed. The sockets
also bind to port 0 so that they will be assigned an open port by the
OS.